// src/app/interfaces/IAttractionTicket.ts
export interface IAttractionTicket {
  fAttractionTicketId?: number | null;
  fAttractionId?: number | null;
  fAttractionName?: string | null;
  fTicketType?: string | null;
  fPrice?: number | null;
  fDiscountInformation?: string | null;
  fCreatedDate?: string | null;
  fimageSrc?: string | null;
}